+++ crates/egui/src/lib.rs
@@ -417,7 +417,7 @@ pub mod response;
 mod sense;
 pub mod style;
 pub mod text_selection;
-pub mod theme_plugin;
+pub mod theme;
 mod ui;
 mod ui_builder;
 mod ui_stack;
